requirements.txt
tkinter
Pillow
markdown2
webbrowser
json
logging
threading
tqdm
fastmcp
re
base64
multiprocessing
time
sys
os
pyinstaller
svglib
reportlab